+++ Plugins/Guardian/Source/NexusGuardian/Private/NGuardianSubsystem.cpp
@@ -6,6 +6,7 @@
 #include "NCoreMinimal.h"
 #include "NGuardianMinimal.h"
 #include "NGuardianSettings.h"
+#include "Async/Async.h"
 #include "Developer/NDeveloperUtils.h"
 #include "Developer/NObjectSnapshot.h"
 #include "Developer/NObjectSnapshotUtils.h"
@@ -61,11 +62,14 @@ void UNGuardianSubsystem::Tick(float DeltaTime)
 		CaptureSnapshot = FNObjectSnapshotUtils::Snapshot();
 		if (bShouldOutputSnapshot)
 		{
-			const FString DumpFilePath = FPaths::Combine(FPaths::ProjectLogDir(),
+			FString DumpFilePath = FPaths::Combine(FPaths::ProjectLogDir(),
 				FString::Printf(TEXT("NEXUS_Snapshot_%s.txt"),*FDateTime::Now().ToString(TEXT("%Y%m%d_%H%M%S"))));
-			FFileHelper::SaveStringToFile(CaptureSnapshot.ToDetailedString(), *DumpFilePath, FFileHelper::EEncodingOptions::ForceUTF8, &IFileManager::Get(), FILEWRITE_Silent);
-			
-			UE_LOG(LogNexusCore, Error, TEXT("A UObject snapshot has been written to %s."), *DumpFilePath);
+			Async(EAsyncExecution::ThreadPool,
+				[Snapshot = CaptureSnapshot, DumpFilePath = MoveTemp(DumpFilePath)]()
+				{
+					FFileHelper::SaveStringToFile(Snapshot.ToDetailedString(), *DumpFilePath, FFileHelper::EEncodingOptions::ForceUTF8, &IFileManager::Get(), FILEWRITE_Silent);
+					UE_LOG(LogNexusCore, Error, TEXT("A UObject snapshot has been written to %s."), *DumpFilePath);
+				});
 		}
 		bPassedObjectCountSnapshotThreshold = true;
 		return;
@@ -79,11 +83,14 @@ void UNGuardianSubsystem::Tick(float DeltaTime)
 		const FNObjectSnapshot CompareSnapshot = FNObjectSnapshotUtils::Snapshot();
 		FNObjectSnapshotDiff Diff = FNObjectSnapshotUtils::Diff(CaptureSnapshot, CompareSnapshot, false);
 
-		const FString DumpFilePath = FPaths::Combine(FPaths::ProjectLogDir(),
+		FString DumpFilePath = FPaths::Combine(FPaths::ProjectLogDir(),
 		FString::Printf(TEXT("NEXUS_Compare_%s.txt"),*FDateTime::Now().ToString(TEXT("%Y%m%d_%H%M%S"))));
-		FFileHelper::SaveStringToFile(Diff.ToDetailedString(), *DumpFilePath, FFileHelper::EEncodingOptions::ForceUTF8, &IFileManager::Get(), FILEWRITE_Silent);
-		
-		UE_LOG(LogNexusGuardian, Error, TEXT("A UObject comparison written to %s."), *DumpFilePath);
+		Async(EAsyncExecution::ThreadPool,
+			[Diff = MoveTemp(Diff), DumpFilePath = MoveTemp(DumpFilePath)]()
+			{
+				FFileHelper::SaveStringToFile(Diff.ToDetailedString(), *DumpFilePath, FFileHelper::EEncodingOptions::ForceUTF8, &IFileManager::Get(), FILEWRITE_Silent);
+				UE_LOG(LogNexusGuardian, Error, TEXT("A UObject comparison written to %s."), *DumpFilePath);
+			});
 		bPassedObjectCountCompareThreshold = true;
 	}
 }